+++ back/src/main/kotlin/ilpak/nomat/NomatApplication.kt
@@ -7,5 +7,5 @@ import org.springframework.boot.runApplication
 class NomatApplication
 
 fun main(args: Array<String>) {
-	runApplication<NomatApplication>(*args)
+    runApplication<NomatApplication>(*args)
 }
