+++ back/src/main/kotlin/ilpak/nomat/exception/NotFoundException.kt
@@ -1,3 +1,3 @@
 package ilpak.nomat.exception
 
-class NotFoundException(message: String): RuntimeException(message)
+class NotFoundException(message: String) : RuntimeException(message)
