+++ back/src/main/kotlin/ilpak/nomat/playlist/dto/PlaylistMetaDataResponse.kt
@@ -1,7 +1,7 @@
 package ilpak.nomat.playlist.dto
 
 data class PlaylistMetaDataResponse(
-	val id: Long,
-	val name: String,
-	val count: Int,
+    val id: Long,
+    val name: String,
+    val count: Int,
 )
