+++ back/src/main/kotlin/ilpak/nomat/room/controller/RoomController.kt
@@ -9,22 +9,22 @@ import org.springframework.web.bind.annotation.*
 @RequestMapping("/rooms")
 @RestController
 class RoomController(
-	private val roomService: RoomService
+    private val roomService: RoomService
 ) {
 
-	@GetMapping
-	fun getRooms(): List<RoomResponse> {
-		return roomService.getRooms()
-	}
+    @GetMapping
+    fun getRooms(): List<RoomResponse> {
+        return roomService.getRooms()
+    }
 
-	@GetMapping("/{roomId}")
-	fun getRoom(@PathVariable roomId: Long): RoomResponse {
-		return roomService.getRoom(roomId)
-	}
+    @GetMapping("/{roomId}")
+    fun getRoom(@PathVariable roomId: Long): RoomResponse {
+        return roomService.getRoom(roomId)
+    }
 
-	@ResponseStatus(HttpStatus.CREATED)
-	@PostMapping
-	fun createRoom(@RequestBody roomRequest: RoomRequest): RoomResponse {
-		return roomService.createRoom(roomRequest)
-	}
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping
+    fun createRoom(@RequestBody roomRequest: RoomRequest): RoomResponse {
+        return roomService.createRoom(roomRequest)
+    }
 }
