+++ back/src/main/kotlin/ilpak/nomat/room/domain/RoomPlaylist.kt
@@ -5,9 +5,9 @@ import jakarta.persistence.Embeddable
 
 @Embeddable
 data class RoomPlaylist(
-	val playlistId: Long,
-	@Column(name = "playlist_name")
-	val name: String,
-	@Column(name = "playlist_count")
-	val count: Int,
+    val playlistId: Long,
+    @Column(name = "playlist_name")
+    val name: String,
+    @Column(name = "playlist_count")
+    val count: Int,
 )
