+++ back/src/main/kotlin/ilpak/nomat/room/dto/PlaylistResponse.kt
@@ -3,17 +3,17 @@ package ilpak.nomat.room.dto
 import ilpak.nomat.room.domain.RoomPlaylist
 
 data class PlaylistResponse(
-	val id: Long,
-	val name: String,
-	val count: Int,
+    val id: Long,
+    val name: String,
+    val count: Int,
 ) {
-	companion object {
-		fun of(playlist: RoomPlaylist): PlaylistResponse {
-			return PlaylistResponse(
-				playlist.playlistId,
-				playlist.name,
-				playlist.count
-			)
-		}
-	}
+    companion object {
+        fun of(playlist: RoomPlaylist): PlaylistResponse {
+            return PlaylistResponse(
+                playlist.playlistId,
+                playlist.name,
+                playlist.count
+            )
+        }
+    }
 }
