+++ back/src/main/kotlin/ilpak/nomat/room/dto/RoomRequest.kt
@@ -1,8 +1,8 @@
 package ilpak.nomat.room.dto
 
 data class RoomRequest(
-	val title: String,
-	val roomCapacity: Int,
-	val password: String?,
-	val playlistId: Long,
+    val title: String,
+    val roomCapacity: Int,
+    val password: String?,
+    val playlistId: Long,
 )
